# Optimize the sha1-mh on aarch64 by SVE/SVE2