--- a/mh_sha1/aarch64/mh_sha1_sve.c
+++ b/mh_sha1/aarch64/mh_sha1_sve.c
@@ -1,5 +1,5 @@
 /**********************************************************************
-  Copyright(c) 2020-2022 Arm Corporation All rights reserved.
+  Copyright(c) 2022 Arm Corporation All rights reserved.
 
   Redistribution and use in source and binary forms, with or without
   modification, are permitted provided that the following conditions
@@ -26,68 +26,28 @@
   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
   OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 **********************************************************************/
-#include <aarch64_multibinary.h>
+#include <string.h>
+#include "mh_sha1_internal.h"
 
-#ifndef HWCAP2_SVE2
-#define HWCAP2_SVE2	(1 << 1)
-#endif
+void mh_sha1_block_sve(const uint8_t * input_data,
+		       uint32_t digests[SHA1_DIGEST_WORDS][HASH_SEGS],
+		       uint8_t frame_buffer[MH_SHA1_BLOCK_SIZE], uint32_t num_blocks);
+/***************mh_sha1_update***********/
+// mh_sha1_update_sve.c
+#define MH_SHA1_UPDATE_FUNCTION mh_sha1_update_sve
+#define MH_SHA1_BLOCK_FUNCTION	mh_sha1_block_sve
+#include "mh_sha1_update_base.c"
+#undef MH_SHA1_UPDATE_FUNCTION
+#undef MH_SHA1_BLOCK_FUNCTION
 
-#define CAP_SVE	1
-#define CAP_SVE2	2
-#define CAP_NOSVE	0
-
-static inline int sve_capable(unsigned long auxval)
-{
-	if (auxval & HWCAP_SVE) {
-		if (getauxval(AT_HWCAP2) & HWCAP2_SVE2) {
-			return CAP_SVE2;
-		}
-		return CAP_SVE;
-	}
-	return CAP_NOSVE;
-}
-
-DEFINE_INTERFACE_DISPATCHER(mh_sha1_update)
-{
-	unsigned long auxval = getauxval(AT_HWCAP);
-
-	if (auxval & HWCAP_SHA1)
-		return PROVIDER_INFO(mh_sha1_update_ce);
-
-	switch (sve_capable(auxval)) {
-	case CAP_SVE:
-		return PROVIDER_INFO(mh_sha1_update_sve);
-	case CAP_SVE2:
-		return PROVIDER_INFO(mh_sha1_update_sve2);
-	default:
-		break;
-	}
-
-	if (auxval & HWCAP_ASIMD)
-		return PROVIDER_INFO(mh_sha1_update_asimd);
-
-	return PROVIDER_BASIC(mh_sha1_update);
-
-}
-
-DEFINE_INTERFACE_DISPATCHER(mh_sha1_finalize)
-{
-	unsigned long auxval = getauxval(AT_HWCAP);
-
-	if (auxval & HWCAP_SHA1)
-		return PROVIDER_INFO(mh_sha1_finalize_ce);
-
-	switch (sve_capable(auxval)) {
-	case CAP_SVE:
-		return PROVIDER_INFO(mh_sha1_finalize_sve);
-	case CAP_SVE2:
-		return PROVIDER_INFO(mh_sha1_finalize_sve2);
-	default:
-		break;
-	}
-
-	if (auxval & HWCAP_ASIMD)
-		return PROVIDER_INFO(mh_sha1_finalize_asimd);
-
-	return PROVIDER_BASIC(mh_sha1_finalize);
-}
+/***************mh_sha1_finalize AND mh_sha1_tail***********/
+// mh_sha1_tail is used to calculate the last incomplete src data block
+// mh_sha1_finalize is a mh_sha1_ctx wrapper of mh_sha1_tail
+// mh_sha1_finalize_sve.c and mh_sha1_tail_sve.c
+#define MH_SHA1_FINALIZE_FUNCTION	mh_sha1_finalize_sve
+#define MH_SHA1_TAIL_FUNCTION		mh_sha1_tail_sve
+#define MH_SHA1_BLOCK_FUNCTION		mh_sha1_block_sve
+#include "mh_sha1_finalize_base.c"
+#undef MH_SHA1_FINALIZE_FUNCTION
+#undef MH_SHA1_TAIL_FUNCTION
+#undef MH_SHA1_BLOCK_FUNCTION