Here are the test results on N2/V1 micro-architecture:

---------+-------------+-------------+------------+
|        | CPU0  (N2)  | CPU2  (N2)  | CPU3  (V1) |
|        | 128-bit VL  | 128-bit VL  | 256-bit VL |
+--------+-------------+-------------+------------+
|  neon  |  923 MB/s   |  1002 MB/s  |  930 MB/s  |
+--------+-------------+-------------+------------+
|  sve2  |  1146 MB/s  |  1634 MB/s  |            |
|        |  (+24%)     |  (+63%)     |            |
+--------+-------------+-------------+------------+
|  sve   |             |             |  1465 MB/s |
|        |             |             |  (+57.5)   |
+--------+-------------+-------------+------------+

Change-Id: Id19f73ec75bd283cca628c0ff516289adb0ac4ff
Signed-off-by: Daniel Hu <Daniel.Hu@arm.com>